+++ CHANGELOG.md
@@ -6,6 +6,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## [2025.10.20]
+
+### Added
+
+- Designed kernel and enhanced interfaces for Playlist component
+
+### Updated
+
+- Changed design to include use of Song record instead of generic Pair<K, V>
+- Added hasSong(), currentArtist(), and currentTitle() secondary methods
+
+## [2025.10.10]
+
+### Added
+
+- Designed a proof of concept for Playlist component
+
+### Updated
+
+- Changed design to include adding playlistSize() kernel method and listAll() secondary method.
+
 ## [2025.09.15]
 
 ### Added